Move dashboard modal to dashboard route to avoid test failures
