+++ ui/app/templates/vault/cluster/dashboard.hbs
@@ -19,3 +19,48 @@
     </Hds::Card::Container>
   {{/if}}
 </div>
+
+{{#if (and this.auth.isActiveSession (not this.shouldHideNewFeaturesModal))}}
+  <Modal
+    @title="New dashboard"
+    @onClose={{action (mut this.showNewFeatureModal) false}}
+    @isActive={{this.showNewFeatureModal}}
+    @showCloseButton={{true}}
+  >
+    <section class="modal-card-body has-top-padding-l has-bottom-padding-l">
+      <div>
+        <div class="has-bottom-margin-s">
+          Vault now has a dashboard that includes, among other things:
+        </div>
+        <ul class="bullet">
+          <li>An overview of Vault's metrics, shortcuts, and at-a-glance information</li>
+          <li>Configuration details such as cluster address, TTL, TLS, log format, and storage type</li>
+          <li>Quick actions: perform important tasks such as finding a secret or generating a leaf certificate</li>
+        </ul>
+        <div class="has-top-margin-s">
+          You can still access the Secrets engines page on the sidebar.
+        </div>
+      </div>
+      <Input
+        id="newDashboardModal"
+        class="has-top-margin-m"
+        @type="checkbox"
+        name="newDashboardModal"
+        @checked={{this.alwaysHideNewFeaturesModal}}
+        {{on "change" this.toggleAlwaysHideNewFeaturesModal}}
+        data-test-do-not-show-me-again
+      />
+      <label for="newDashboardLabel">Don’t show me this again</label>
+    </section>
+    <footer class="modal-card-foot modal-card-foot-outlined">
+      <button
+        type="button"
+        class="button is-primary"
+        {{on "click" (fn (mut this.showNewFeatureModal) false)}}
+        data-test-new-features-modal-button
+      >
+        Got it
+      </button>
+    </footer>
+  </Modal>
+{{/if}}
